requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
cached-property==1.5.2
celery==5.1.2
celery-progress==0.1.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-celery-results==2.2.0
django-crispy-forms==1.12.0
django-import-export==2.5.0
et-xmlfile==1.1.0
fake-useragent==0.1.11
idna==2.10
importlib-metadata==2.1.1
install==1.3.4
kombu==5.1.0
lxml==4.6.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
parse==1.19.0
pkg-resources==0.0.0
prompt-toolkit==3.0.19
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==8.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
